separate PipelineWorkflow & ResourceReference methods from pipeline.go
